+++ src/common/constants/api.js
@@ -2,3 +2,6 @@ export const TITLES_API = 'orders/titles';
 export const TITLE_API = `${TITLES_API}/:{id}`;
 export const PO_LINES_API = 'orders/order-lines';
 export const LOCATIONS_API = 'locations';
+export const PIECES_API = 'orders-storage/pieces';
+export const ORDER_PIECES_API = 'orders/pieces';
+export const CHECKIN_API = 'orders/check-in';