+++ src/common/resources/locationResources.js
@@ -9,6 +9,9 @@ import {
 export const locationsResource = {
   ...baseManifest,
   path: LOCATIONS_API,
+  params: {
+    query: 'cql.allRecords=1 sortby name',
+  },
   accumulate: true,
   records: 'locations',
 };